UNI-187 #comment Refactor Error classes a bit. Fix mystery test. Add dox.
