+++ unicorn/frontend/lib/UserError.js
@@ -19,15 +19,87 @@
 
 'use strict';
 
-//
-//Overcomes the fact babel have problems inheriting from builtin objects
-//FIXME: See http://stackoverflow.com/questions/3108980
-//
+
+/**
+ * Base UserError class
+ *
+ * Overcomes the fact babel have problems inheriting from builtin objects:
+ *  @TODO: http://stackoverflow.com/questions/3108980
+ *
+ * @class
+ * @extends Error
+ * @module
+ * @this UserError
+ */
 export default class UserError extends Error {
+  /**
+   * @constructor
+   * @default
+   * @method
+   * @param {string} message - Error message for user
+   * @this UserError
+   */
   constructor(message) {
     super(message);
     this.message = message;
     this.stack = (new Error()).stack;
     this.name = this.constructor.name;
+  }
+}
+
+/**
+ * Thrown when having trouble getting data from Database
+ * @class
+ * @extends UserError
+ * @module
+ * @this DatabaseGetError
+ */
+export class DatabaseGetError extends UserError {
+  /**
+   * @constructor
+   * @method
+   * @param {string} message - Error message for user
+   * @this DatabaseGetError
+   */
+  constructor(message) {
+    super(message || 'Could not perform GET operation from Database');
+  }
+}
+
+/**
+ * Thrown when having trouble putting data into Database
+ * @class
+ * @extends UserError
+ * @module
+ * @this DatabasePutError
+ */
+export class DatabasePutError extends UserError {
+  /**
+   * @constructor
+   * @method
+   * @param {string} message - Error message for user
+   * @this DatabasePutError
+   */
+  constructor(message) {
+    super(message || 'Could not perform PUT operation into Database');
+  }
+}
+
+/**
+ * Thrown when having trouble getting data from Filesystem
+ * @class
+ * @extends UserError
+ * @module
+ * @this FilesystemGetError
+ */
+export class FilesystemGetError extends UserError {
+  /**
+   * @constructor
+   * @method
+   * @param {string} message - Error message for user
+   * @this FilesystemGetError
+   */
+  constructor(message) {
+    super(message || 'Could not perform READ operation on Filesystem');
   }
 }
